+++ glean-core/android/src/main/java/mozilla/telemetry/glean/debug/GleanDebugActivity.kt
@@ -84,6 +84,7 @@ class GleanDebugActivity : Activity() {
             return
         }
 
+        // Make sure that at least one of the supported commands was used.
         val supportedCommands = listOf(
             SEND_PING_EXTRA_KEY,
             LOG_PINGS_EXTRA_KEY,
@@ -94,6 +95,7 @@ class GleanDebugActivity : Activity() {
 
         var nextActivityName: String? = null
 
+        // Enable debugging options and start the application.
         intent.extras?.let {
             it.keySet().forEach { cmd ->
                 if (!supportedCommands.contains(cmd)) {
@@ -101,25 +103,49 @@ class GleanDebugActivity : Activity() {
                 }
             }
 
-            intent.getStringExtra(TAG_DEBUG_VIEW_EXTRA_KEY)?.let { Glean.setDebugViewTag(it) }
-            intent.getBooleanExtra(LOG_PINGS_EXTRA_KEY, false).let { Glean.setLogPings(it) }
-            intent.getStringArrayExtra(SOURCE_TAGS_KEY)?.let { tags -> Glean.setSourceTags(tags.toSet()) }
-            intent.getStringExtra(NEXT_ACTIVITY_TO_RUN)?.let { nextActivityName = it }
+            // Check for ping debug view tag to apply to the X-Debug-ID header when uploading the
+            // ping to the endpoint
+            val debugViewTag: String? = intent.getStringExtra(TAG_DEBUG_VIEW_EXTRA_KEY)
+
+            // Set the debug view tag, if the tag is invalid it won't be set
+            debugViewTag?.let {
+                Glean.setDebugViewTag(debugViewTag)
+            }
+
+            val logPings: Boolean? = intent.getBooleanExtra(LOG_PINGS_EXTRA_KEY, false)
+            logPings?.let {
+                Glean.setLogPings(logPings)
+            }
+
+            intent.getStringArrayExtra(SOURCE_TAGS_KEY)?.let { tags ->
+                Glean.setSourceTags(tags.toSet())
+            }
+
+            intent.getStringExtra(NEXT_ACTIVITY_TO_RUN)?.let { nextActivity ->
+                nextActivityName = nextActivity
+            }
+
             // Important: this should be applied as the last one, so that
             // any other option will affect the ping submission as well.
-            intent.getStringExtra(SEND_PING_EXTRA_KEY)?.let { Glean.submitPingByName(it) }
+            intent.getStringExtra(SEND_PING_EXTRA_KEY)?.let { name ->
+                Glean.submitPingByName(name)
+            }
         }
 
         // This Activity can be used to tag tests on CI or start products with specific
         // options. We need to make sure to retain and propagate all the options that
         // we were passed to the next intent. Our 3 steps strategy:
-        // 1. Copy intent and make extras immutable
-        val safeExtras = Bundle(intent.extras).apply { setClassLoader(javaClass.classLoader) }
-        val nextIntent = Intent().apply { replaceExtras(safeExtras) }
+        // 1. use the `Intent` copy constructor to copy all the intent options from the
+        //   intent which started this activity;
+        val nextIntent = Intent(intent)
 
-        // 2. Determine the next component
+        // 2. get the main launch intent for the product using the Glean SDK or lookup
+        //    the one passed as a command line argument to this activity.
         val launchIntent = packageManager.getLaunchIntentForPackage(packageName)!!
+        nextIntent.`package` = launchIntent.`package`
+
         val nextComponent = nextActivityName?.let { name ->
+            // Try to lookup the Activity that was asked by the caller.
             val component = ComponentName(packageName, name)
             if (!isActivityExported(component)) {
                 Log.e(LOG_TAG, "Cannot run $packageName/$name: Activity not exported")
@@ -127,17 +153,14 @@ class GleanDebugActivity : Activity() {
                 return
             }
             component
-        } ?: launchIntent.component!!
+        } ?: launchIntent.component
 
-        Log.i(LOG_TAG, "Running next: ${nextComponent.packageName}/${nextComponent.className}")
+        Log.i(LOG_TAG, "Running next: ${nextComponent!!.packageName}/${nextComponent.className}")
 
-        // 3. Configure and launch intent safely
-        nextIntent.apply {
-            setClassName(packageName, nextComponent.className)
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
-        }
-
+        // 3. change the starting "component" to the one from the previous step.
+        nextIntent.component = nextComponent
         startActivity(nextIntent)
+
         finish()
     }
 }