+++ glean-core/src/lib.rs
@@ -59,7 +59,7 @@ mod fd_logger;
 
 pub use crate::common_metric_data::{CommonMetricData, Lifetime};
 pub use crate::core::Glean;
-pub use crate::core_metrics::ClientInfoMetrics;
+pub use crate::core_metrics::{AttributionMetrics, ClientInfoMetrics, DistributionMetrics};
 pub use crate::error::{Error, ErrorKind, Result};
 pub use crate::error_recording::{test_get_num_recorded_errors, ErrorType};
 pub use crate::histogram::HistogramType;
@@ -101,6 +101,10 @@ static PRE_INIT_SOURCE_TAGS: Mutex<Vec<String>> = Mutex::new(Vec::new());
 /// Keep track of pings registered before Glean is initialized.
 static PRE_INIT_PING_REGISTRATION: Mutex<Vec<metrics::PingType>> = Mutex::new(Vec::new());
 static PRE_INIT_PING_ENABLED: Mutex<Vec<(metrics::PingType, bool)>> = Mutex::new(Vec::new());
+
+/// Keep track of attribution and distribution supplied before Glean is initialized.
+static PRE_INIT_ATTRIBUTION: Mutex<Option<AttributionMetrics>> = Mutex::new(None);
+static PRE_INIT_DISTRIBUTION: Mutex<Option<DistributionMetrics>> = Mutex::new(None);
 
 /// Global singleton of the handles of the glean.init threads.
 /// For joining. For tests.
@@ -446,6 +450,15 @@ fn initialize_inner(
                 let pings = PRE_INIT_PING_ENABLED.lock().unwrap();
                 for (ping, enabled) in pings.iter() {
                     glean.set_ping_enabled(ping, *enabled);
+                }
+
+                // The attribution and distribution might have been set before initialize,
+                // take the cached values and set them.
+                if let Some(attribution) = PRE_INIT_ATTRIBUTION.lock().unwrap().take() {
+                    glean.update_attribution(attribution);
+                }
+                if let Some(distribution) = PRE_INIT_DISTRIBUTION.lock().unwrap().take() {
+                    glean.update_distribution(distribution);
                 }
 
                 // If this is the first time ever the Glean SDK runs, make sure to set
@@ -1255,6 +1268,50 @@ pub fn glean_set_dirty_flag(new_value: bool) {
     core::with_glean(|glean| glean.set_dirty_flag(new_value))
 }
 
+/// Updates attribution fields with new values.
+/// AttributionMetrics fields with `None` values will not overwrite older values.
+pub fn glean_update_attribution(attribution: AttributionMetrics) {
+    if was_initialize_called() {
+        core::with_glean(|glean| glean.update_attribution(attribution));
+    } else {
+        PRE_INIT_ATTRIBUTION
+            .lock()
+            .unwrap()
+            .get_or_insert(Default::default())
+            .update(attribution);
+    }
+}
+
+/// **TEST-ONLY Method**
+///
+/// Returns the current attribution metrics.
+/// Panics if called before init.
+pub fn glean_test_get_attribution() -> AttributionMetrics {
+    core::with_glean(|glean| glean.test_get_attribution())
+}
+
+/// Updates distribution fields with new values.
+/// DistributionMetrics fields with `None` values will not overwrite older values.
+pub fn glean_update_distribution(distribution: DistributionMetrics) {
+    if was_initialize_called() {
+        core::with_glean(|glean| glean.update_distribution(distribution));
+    } else {
+        PRE_INIT_DISTRIBUTION
+            .lock()
+            .unwrap()
+            .get_or_insert(Default::default())
+            .update(distribution);
+    }
+}
+
+/// **TEST-ONLY Method**
+///
+/// Returns the current distribution metrics.
+/// Panics if called before init.
+pub fn glean_test_get_distribution() -> DistributionMetrics {
+    core::with_glean(|glean| glean.test_get_distribution())
+}
+
 #[cfg(all(not(target_os = "android"), not(target_os = "ios")))]
 static FD_LOGGER: OnceCell<fd_logger::FdLogger> = OnceCell::new();
 
@@ -1307,31 +1364,19 @@ mod ffi {
 
     type CowString = Cow<'static, str>;
 
-    impl UniffiCustomTypeConverter for CowString {
-        type Builtin = String;
-
-        fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-            Ok(Cow::from(val))
-        }
-
-        fn from_custom(obj: Self) -> Self::Builtin {
-            obj.into_owned()
-        }
-    }
+    uniffi::custom_type!(CowString, String, {
+        remote,
+        lower: |s| s.into_owned(),
+        try_lift: |s| Ok(Cow::from(s))
+    });
 
     type JsonValue = serde_json::Value;
 
-    impl UniffiCustomTypeConverter for JsonValue {
-        type Builtin = String;
-
-        fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-            Ok(serde_json::from_str(&val)?)
-        }
-
-        fn from_custom(obj: Self) -> Self::Builtin {
-            serde_json::to_string(&obj).unwrap()
-        }
-    }
+    uniffi::custom_type!(JsonValue, String, {
+        remote,
+        lower: |s| serde_json::to_string(&s).unwrap(),
+        try_lift: |s| Ok(serde_json::from_str(&s)?)
+    });
 }
 pub use ffi::*;